requirements.txt
# requirements.txt  (excerpt)
python-decouple==3.8        
Django==5.1.3

django-parler==2.3
gunicorn==23.0.0

Markdown==3.8